requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
black==24.8.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coverage==7.6.1
Django==4.2.2
django-cors-headers==4.4.0
django-filter
djangorestframework==3.15.2
djangorestframework-simplejwt
dnspython==2.6.1
drf-yasg==1.21.7
greenlet==3.0.3
inflection==0.5.1
kombu==5.4.0
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
prompt_toolkit==3.0.47
psycopg2==2.9.9
psycopg2-binary==2.9.9
PyJWT==2.9.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
setuptools==74.1.2
six==1.16.0
sqlparse==0.5.1
tzdata==2024.1
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.13
python-dotenv~=1.0.1